+++ examples/toroid-repl/term_cbreak_darwin.go
@@ -1,0 +1,9 @@
+//go:build darwin
+// +build darwin
+
+package main
+
+import "golang.org/x/sys/unix"
+
+func getTermios() uint { return unix.TIOCGETA }
+func setTermios() uint { return unix.TIOCSETA }